refactor(#244): 강사 신고내역 컴포넌트 분리
