+++ src/app/mypage/instructor/report/page.tsx
@@ -1,21 +1,19 @@
 'use client';
-import { useEffect, useState } from 'react';
-import { NoteSVG } from '@/icons/svg';
+import { useQuery } from '@tanstack/react-query';
+import dynamic from 'next/dynamic';
 import { getInstructorReport } from '@/lib/apis/reportApis';
-import ReportModal from './ReportModal';
-import { IReportResponse } from '@/types/report';
+
+const ReportList = dynamic(() => import('./_components/ReportList'), {
+  ssr: false,
+});
 
 const ReportHistoryPage = () => {
-  const [userReportData, setUserReportData] = useState<IReportResponse[]>([]);
+  const { data, isLoading, error } = useQuery({
+    queryKey: ['instructor', 'report'],
+    queryFn: () => getInstructorReport(100, 0, 0, 0, 0, 'ALL'),
+  });
 
-  useEffect(() => {
-    const fetchData = async () => {
-      const reportData = await getInstructorReport(100, 0, 0, 0, 0, 'ALL');
-      setUserReportData(reportData.reportList);
-    };
-
-    fetchData();
-  }, []);
+  if (!data || error) return null;
 
   return (
     <section className="mx-4 box-border flex flex-col rounded-lg bg-white px-3.5 py-5 text-sm text-gray-100 md:mx-9 md:px-5">
@@ -30,8 +28,8 @@ const ReportHistoryPage = () => {
           </tr>
         </thead>
         <tbody>
-          {userReportData.map((item) => (
-            <TableList key={item.id} {...item} />
+          {data.map((item) => (
+            <ReportList key={item.id} {...item} />
           ))}
         </tbody>
       </table>
@@ -40,71 +38,3 @@ const ReportHistoryPage = () => {
 };
 
 export default ReportHistoryPage;
-
-const TableList = ({
-  targetUser,
-  targetLecturer,
-  reason,
-  userReportType,
-  isAnswered,
-  userReportResponse,
-}: IReportResponse) => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const openModal = () => {
-    setIsModalOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-  };
-
-  const target = targetUser
-    ? targetUser.nickname
-    : targetLecturer
-    ? targetLecturer.nickname
-    : null;
-  const status = isAnswered ? '처리완료' : '처리중';
-
-  const reportTypes = userReportType.map((type) => type.reportType.description);
-  const str = reportTypes.length > 1 ? ` 외 ${reportTypes.length - 1}` : '';
-
-  const response = userReportResponse?.description;
-
-  return (
-    <tr className="flex items-start gap-4 border-b border-solid border-gray-700 px-4 text-left font-medium md:gap-10">
-      <th className="h-full w-20 py-2 md:w-24">{target}</th>
-      <th className="flex w-32 flex-1 flex-col items-center gap-[0.31rem] py-2 md:w-36">
-        <p className="flex w-full items-center">
-          {reportTypes[0] + str}
-
-          <NoteSVG
-            width="16"
-            height="16"
-            onClick={openModal}
-            className="ml-1.5 cursor-pointer stroke-gray-500 hover:stroke-black"
-          />
-        </p>
-        <div>
-          <ReportModal
-            isOpen={isModalOpen}
-            closeModal={closeModal}
-            target={target}
-            reportTypes={reportTypes}
-            reason={reason}
-            status={status}
-            response={response}
-          />
-        </div>
-      </th>
-      <th
-        onClick={openModal}
-        className={`w-12 cursor-pointer whitespace-nowrap py-2 ${
-          status === '처리중' ? 'text-sub-color1' : 'text-gray-500'
-        } underline`}
-      >
-        {status}
-      </th>
-    </tr>
-  );
-};
